Minor style updates
I think the glob import was good to start with, but I think things are
stable enough to let rust-analyzer auto-expand it.

I'm not sure there was ever a good reason to use `NonNull` on the YAML
parser error initialization, especially with that *const to *mut cast in
there.

Finally, I absolutely do think there are cases where a match on a
boolean can be more readable than an if expression, but I don't think
the case I touched here is one of them.