Restore version and help output
This version of xt's help flags still maintains a distinction between
short and long help output, but now uses an *especially* short help
output for the '-h' flag that just prints basic usage information.

I've used a variety of other CLI tools as inspiration for various
aspects of the presentation. In particular, the formatting of the usage
strings looks somewhat like `iconv`, and the presentation of prose
documentation after basic usage information is somewhat like `jq`.
Honestly, my favorite part is that the options and flags in the `--help`
output aren't blown out onto multiple lines, which pushes a whole lot of
useful information off of a lot of reasonably-sized terminals.

I've also rewritten parts of the prose to better clarify things like
multi-document support and the special '-' file name for standard input,
and split up the detailed information for each format onto different
lines for easier reading.

A notable concern at the moment is that the program name is always
printed as "xt" in usage strings, even if that's not actually the name
with which the program was executed. I can probably fix that with a bit
of work, even though I don't think it's *that* big of a deal.