Lock the handle for standard input
Apparently every read on the non-locked `Stdin` just gets this same lock
right before it executes the read, so we might as well take it once.

I've been thinking for a while that I should re-work xt's reader handles
to expect `BufRead` rather than `Read`. I guess if the only safe way to
read stdin is through this special buffered proxy, that just strengthens
the argument.