Eliminate all remaining eprintln!() calls
eprintln!() panics on write errors, which in our case means an abort,
which we want as little chance of as possible. I mean, it's silly to
turn an attempt to write a regular error message into a core dump just
because the stderr pipe broke or something, right?

Naturally we ignore any errors in writing the output, since if we can't
even print an error message and we don't want to panic it's not like we
have many more options.