Warn on all uses of [e]print[ln]!()
I intentionally got rid of all of these, and didn't realize until now
that there was a lint to catch their usage.

To be clear, it is both short-sighted and morally wrong to block the use
of temporary prints for debugging, so this should NEVER be a hard deny.
But it may well be the only set of lints for which I'd say that's the
case.