Update comments around mmap usage
Where the previous comment talked about using memory maps "in the name
of performance," I feel like it left out a lot of important nuance, and
possibly gave the impression that I have not properly considered the
problems and tradeoffs associated with memory maps.

As such, I want to clarify that the "performance" aspect of memory maps
in xt is all about facilitating the use of slice-based input rather than
reader-based input, and that reading a file into memory normally would
almost certainly provide equivalent user-facing performance. The true
benefit of a memory map, in my mind, is that we can obtain the benefits
of slice-based input while giving the OS a useful hint that the contents
of this memory are ultimately based on the contents of a file, which
might better facilitate the OS deciding to reclaim this memory for more
important purposes, especially if it does so in preference to swapping
out anonymous pages. Of course, we also take on the possibility of
undefined program behavior in the event that the backing file changes,
and we can only hope that we get a SIGBUS upon the file being truncated
rather than something more terrifying. So, again, not a magical feature
with no tradeoffs.