Deprecate public functions using `input::Handle`
After giving some thought to xt's library API, I strongly feel that
`Handle` should remain an internal implementation detail rather than
something that arbitrary clients can instantiate, and that specific
slice and reader translation functions make code more readable than when
a handle has to be constructed and passed.

This is a non-trivial change, since there are a lot of different parts
of xt that assume they can create a handle and pass it to a translation
function later. So, I'm starting out by marking the existing functions
as deprecated, which should make them easy to identify and replace in
future commits.