Fix builds on Linux
This is EXTREMELY embarrassing... I've only been trying to build the
program on macOS, so the latest releases with streaming YAML support
simply do not build AT ALL on Linux.

I don't even have words to describe what a piece of $#@! I am, having
fully tagged and released something like this to the public, and looking
at changelogs of the last few days and seeing nothing but inordinate
amounts of bikeshedding as potential users can't even try the program. I
have yet to decide what's going to happen to the broken releases on
crates.io. The least bad option may simply be to release v0.15.0 as-is,
and then NEVER release again until some semblance of proper automated
testing is in place.