Fix build failure on Rust 1.59.0
Once again, my shoddy development and release practices have led to
another embarrassing hotfix and crates.io yank, as I have learned that
my code does not build properly on the MSRV documented in Cargo.toml. It
is obviously NOT acceptable to bump the MSRV in a non-MINOR (since this
is v0.x) release, therefore I've reworked the code a tiny bit to
successfully compile. (Notably, the usages of stdio locks in the
xt_fail! macro don't seem to exhibit the issue.)