Add an initial "full test matrix" manual workflow
This new workflow tests xt on Linux, macOS, and Windows with both stable
Rust and the MSRV, and performs an additional test run using Miri on
nightly. After making sure that this is reasonably stable, I'll almost
certainly add a cut-down version to run on every push to the main
branch, since I don't want to use up tons of resources for nearly every
single commit.

Even with just the manual trigger, this is a major improvement in
testing capabilities for xt compared to the previous MacBook-only
strategy that led to yesterday's shameful and embarrassing series of
releases and crates.io yanks.