Don't fail fast on matrix testing
With the way the matrix is set up, we shouldn't necessarily assume that
a failure in one job correlates with failures in other jobs.

In this case, multiple parts of my first run attempt were canceled due
to a CRLF issue in the integration tests on Windows, which killed the
macOS jobs even though their results would have been useful.