Attempt to fix integration tests on Windows
Since it just seems to be the on-disk versions of these files that have
the CRLF issue, it sounds like Git on Windows (at least in GitHub's
setup) is converting the line endings by default. This seems to be the
recommended approach to control that behavior on a per-repo basis (so it
works for everyone, not just inside of the GitHub Actions setup).