Eliminate an improper Deref impl
It is not acceptable to implement Deref for anything other than smart
pointers. This is obvious from the documentation and called out in
community resources that I have clearly not paid enough attention to.

Writing a custom accessor method feels to me like a more reasonable
option than taking the risk of misusing yet another trait.