Fix potential generation of a blank version string
Since the concatenated version string always includes a space, it could
never possibly be empty.

Do I even need to try to be this friendly to non-Cargo builds? Since I
apparently can't even spot *this* obvious of a bug on the first pass,
it's not like I deserve to say this version is more right than the last
one. All I know right now is that the output on my own system using
Cargo doesn't seem to be completely broken.