Migrate from atty to is-terminal
I guess all the cool people are doing this due to a soundness issue on
Windows that doesn't actually cause problems in practice but hasn't been
fixed for a long time. This adds a ton of dependencies, but most are
Microsoft's official Rust libraries for Windows on various platforms.
The binary size on macOS is actually around 1 kB smaller.

I'm also using a locked stdout handle for output now, for whatever that
could possibly be worth.