Upgrade most dependencies
The winnow crate was kept at v0.4.1, as v0.4.2 bumped the MSRV from
1.60.0 to 1.64.0, and I don't have any particular reason to bump xt's
own MSRV right now. winnow is a transitive dependency of the toml crate
that explicitly only guarantees support for Rust releases within the
last 6 months, so this may unfortunately become something of a common
occurrence.

Anyways, we're now using the latest version of crossbeam-channel (a
transitive dependency of criterion), so there should be no more warnings
about a yanked crate while installing the xt binary (even though that
particular crate is not actually a dependency of the xt binary).