Flatten a nested if-else
Last night I had the "bright idea" to remove a level of nesting by doing
a single `if matches!(input, Input::Stdin) && stdin_used` followed by an
`else` to set `stdin_used = true`. Of course this is wrong since it sets
`stdin_used = true` for *all* inputs, not just stdin, and thankfully I
realized that before I committed it.

I think at one point I had this idea that the nested else actually
looked nicer due to both the xt_bail and the assignment being at the
same level of nesting... but I think it's time to admit that I was
clearly wrong on this, and that the form in which the assignment always
runs makes it way more obvious that my apparent "cleanup" was logically
incorrect.