Bump MSRV to Rust 1.70
Despite being just over a year old, Rust 1.61 is too old to support
several dependency upgrades, including a criterion upgrade that will
fully remove all references to the memory-unsafe (in extremely rare
situations that don't actually apply to xt) atty crate.

There are two reasons I'm selecting this particular version, even though
it's a fairly large bump. First, it's one version behind the latest
stable release, which means it's had 6 full weeks of real-world usage as
a stable release. Second, it introduces the `IsTerminal` trait in the
standard library, which will enable the removal of the third-party
`is-terminal` crate. There are some other side benefits, like support
for the sparse crates.io registry protocol, that are not unique to 1.70
but are nonetheless welcome. Also, I hope that this large of a bump now
will forestall the next MSRV-related breaking change for a bit longer.