Initial fix for YAML chunker unsoundness with broken readers
There was no excuse for this error whatsoever. I discovered it on my own
while I was polishing up the SAFETY comment (since "valid buffer" was a
terrible choice of wording, given that we are assuming the memory may be
uninitialized), but upon further review I've noticed that this condition
is explicitly called out in the documentation for Read::read.

While the only Read impl that's explicitly passed to a Chunker is
Encoder, an Encoder can pass through read operations directly to the
underlying reader in the case of UTF-8 input. This means that even
though Chunker itself is not part of xt's library interface, IT IS
LIKELY POSSIBLE FOR ARBITRARY READERS GIVEN TO XT TO INVOKE THE
UNSOUNDNESS when xt is used as a library. The CLI is not affected to my
knowledge, however it is clear that my knowledge needs to be treated as
suspect.

Once I have tested this change on all platforms through GitHub actions,
I plan to cut a new release of xt and to yank all un-yanked versions
since v0.14.0 (in which streaming YAML support was introduced). The
change itself is frankly not very good, and needs some additional
cleanup and unit tests. However, I'd rather start with something small
that I'm reasonably confident will patch the hole, so that I can spend
more time going very deep on the cleanup without feeling bad about
continuing to have this critical issue released to the public.

As of this writing, I am hesitant to fully remove the YAML streaming
support that the Chunker enables. However, given the damage that unsound
code does to the broader Rust ecosystem, this may be worth serious
consideration. This is not the first time that I have introduced
unsoundness into xt through the YAML chunker, and while the previous
unsoundness was thankfully identified and eliminated before it was
merged into the mainline and released, I think it is becoming
increasingly clear that I cannot be trusted to write unsafe Rust in any
meaningful capacity. (At the same time, I realize that this is a fairly
heavy thing to think about for a project that is practically unused by
the objective metrics that I'm aware of, and that discovering my
inaptitude for unsafe Rust in a small project like this is preferable to
*not* discovering it at all, and as a result convincing myself or others
that I actually "know" Rust, and as a result bringing my faulty
"knowledge" into a context in which I could cause significantly greater
harm.)