Review and update SAFETY comments in the YAML chunker
Given the discovery of a buffer overflow vulnerability in a previous
version of the chunker, I think it's necessary to re-analyze the entire
module nearly from scratch and provide significantly more detail around
various expectations.

In this initial pass, I've rewritten every SAFETY comment from scratch,
but have only made relatively small and obvious changes to the actual
code. I think that the comments require further editing (especially to
tighten them up), and there are some other things I'll need to explore
in terms of the implementation, but I think this is a better starting
point for that work than what was in here before.