Add a unit test for chunking a misbehaving reader
Upon looking at the results of this test, it turns out that xt was
probably not vulnerable to any buffer overflows in practice as a result
of the forgotten length check in the libyaml read handler.

Every reader provided to a Chunker is wrapped by an internal
ChunkReader, which copies the bytes produced by the original reader into
its own internal buffer. Since the buffer we provide to the ChunkReader
is constrained to the size of the buffer provided by libyaml, and since
the ChunkReader's access to that buffer is done entirely in safe code,
it actually performs the necessary bounds check to prevent us from later
overflowing the libyaml buffer, by panicking when creating the slice
that we pass to `extend_from_slice`.

I'm not sure that I actually want to un-yank the old versions, but I'll
definitely update the CHANGELOG notes to account for this.

I'm also noticing that I seem to have dropped the note about why we
maintain our own buffer in the first place? I'll have to bring that
back; it's quite ridiculous that I lost it while ostensibly making the
comments better.