Restore yanked versions to crates.io
As explained in the last commit, the safe ChunkReader implementation
protected us in practice from the potential buffer overflow in the
unsafe libyaml read handler. This is obviously NOT something to depend
on long-term, but I don't see any reason to keep these versions yanked
when there is no specific known vulnerability within them (especially
since v0.17 and below are compatible with older Rust versions).