Re-add the note about why the chunker has its own buffer
Also, remove some of the more overly pedantic elements of the alignment
portion of the copy_nonoverlapping SAFETY comment to tighten it up a
bit. Nobody except for intentional trolls or people making unfunny jokes
is actually going to argue that I should care about byte sizes other
than 8 bits, or that a "multiple of the alignment" could be any
arbitrary real number.