Perform a checked UTF-8 conversion in the YAML chunker
In practice, these UTF-8 validity checks are just not that expensive.
Every time I try to "optimize" them, the real-world difference is just
not that noticeable compared to the work of actually parsing and
transcoding the input.

In the particular case of the YAML chunker, the safety of the UTF-8
conversion assumes more than what was documented about libyaml. As
vaguely mentioned in a TODO comment, it also relies on the correctness
of the ChunkReader, so that it won't e.g. have an off-by-one error that
splits multi-byte characters in bad spots. Instead of writing that up,
I'd rather just get rid of the unsafety until we can prove that it makes
a more meaningful user-facing difference.