Resolve the last clippy warnings
I finally got my VS Code using the new `overrideCommand` setting for the
rust-analyzer extension, and applied the cleanup on the pattern match.
I'm still not sure what was going on with the "items after test module"
issue in the integration tests, so I'm just allowing that one. (It might
not even be under my control, depending on how rstest works.)