Yet another pass over the transcoder docs
I noticed a couple of comments that, after taking a good break from this
module, felt like they leaned too far in the direction of "what it does"
rather than "_why_ it does."

This is probably the least-extensive cleanup I've wanted to do on this
module since I wrote it, which feels like a good sign. I think the only
remaining "weirdness" is that I've always found full match statements
more clear and readable than the use of combinators like `map_err`, even
in the presence of `Ok(x) => Ok(x)`-style stuttering. I'm sure I would
get dinged for that if I asked someone with actual Rust experience for a
code review, but whenever I try it (at least in the context of this
module) I never feel great about the result.