Make the truly CPU-burning options optional
Because xt is such a small binary regardless of how it's built, I've
often been _personally_ willing to accept the tradeoff of waiting a few
more seconds and consuming more energy on a release build to shave the
size of the binary and obtain a small but measurable improvement in
artificial benchmarks. Until now, though, I wasn't ashamed enough of
myself _forcing_ others to accept that same tradeoff to figure out how
to fix it using the custom profiles feature that's been stable since
Cargo 1.57.

(To be clear, I _have_ been ashamed of myself, and by "others" accepting
this tradeoff I mean that I'm also sick of waiting 2 - 3 minutes for a
build on my little arm64 SBC to finish just because I have no problem
using these options on a MacBook Pro and _also_ haven't found motivation
to set up reliable cross-compilation, but somehow I assumed this feature
was still nightly-only.)

Just as importantly: the commit that introduced these options provided a
small amount of context around testing I'd done in my particular
environment at the time to justify their inclusion, but there was no
inline comment about this, so I'm sure it looked like I was
thoughtlessly cargo-culting a semi-random combination of "make it fast"
options. I've tried to add greater peer pressure on users of this option
(yes, including myself) to think more carefully before selecting this
non-default build configuration.

(I hope there's at least a slight hint that my use of `panic = "abort"`
isn't just intended as a cargo-culted "make it fast" option, in the fact
that I've also set it in the development profile. I'm sure this is an
extreme and barely justifiable opinion, but I basically consider stack
unwinding an anti-pattern in languages that provide better mechanisms
for regular error handling, and I do actively want my Rust software to
crash as hard as possible on any form of panic in any context, even when
I'm filling up my development machine with core dumps. Maybe that's also
a lot to push, though, especially since Rust's standard unit testing
framework absolutely requires unwinding panics, which means that the dev
and test configurations aren't 100% identical. But is it really the
right alternative to not keep dev and release equivalent in that
respect? Or to accept in every possible build configuration the cost of
a feature I don't believe in using?)