Stop using compressed input for benchmarks
Based on the recent reveal of the xz-utils backdoor that was obfuscated
in part as compressed test data, I've chosen to eliminate my personal
use of non-human-readable inputs for the xt benchmarks.

As part of this, I don't just want to turn the large json-iterator data
into a JSON file more than 20 MB in size, but instead want to reconsider
how the data for the large benchmark is constructed in the first place,
since an input that large is no better than an input obfuscated via
compression. I'm removing the large benchmarks while I think through
that (which I'm not treating as a major priority, for the record).