Add --help and --version to the man page
This follows the mdoc(7) recommendation to list alphabetically, rather
than keeping the options vs. flags split I have in the current help
output. This is obviously a good recommendation for a command with a lot
of options, but I was surprised to find I prefer it even for the four
options xt has.

At some point I may update the built-in help output to better resemble
the conventions in the man page.