Try taking a fresh look at the YAML chunker
The original chunker was one of my first forays into writing a serious
amount of unsafe code. Being my first such attempt, two things happened:

1. Being unfamiliar with multiple aspects of unsafe code (e.g. the
   "boundaries" between references and pointers), I often took something
   of a "guess and check" approach. For example, I realized I could
   solve Stacked Borrows violations in Miri by using raw pointers, so I
   spread a _lot_ of raw pointers around the module (e.g. the function
   to create a new Event).

2. When I _almost_ had a memory safety disaster due to a forgotten
   boundary check before a memcpy, I "solved" the problem by writing
   ridiculously detailed safety comments covering every aspect I could
   think of, but not by refactoring in any meaningful way.

Trying to prove the 1-byte alignment of a u8 pointer from first
principles might be overkill, but answering the question of "is this
_really_ aligned?" was a meaningful learning exercise that helped me
reach this very point, where I can step back and see this all from a
higher level. I mentioned that Event::from_parser is weird for taking a
raw parser pointer, but I think the overall pattern of wrapping it as a
smart pointer type--where I can mostly use it as a reference and rely on
&self in Deref to enforce the right aliasing rules--is worth emulating.

My first foray with unsafe was about taking the advice to "make unsafe
blocks small" and applying it as literally as possible, to the point
that common functions were getting paragraph-long safety justifications
focused on basic properties of a Box-derived raw pointer. I had to keep
asking myself: how do I know all of this is correct?

Now, I'm asking a new question: how do I build abstractions that
_protect_ the invariants required to guarantee safety? And I think the
key is to scope those invariants as tightly as possible by leaning into
this smart pointer idea. I don't want to ensure that a pointer is valid
because I kept it valid all the way down a call stack; I want it to be
valid because it's permanently locked behind Deref[Mut], and I checked
my work in there.

I also want to remove panics where it makes sense, given the reality
that unsafe_libyaml is no longer maintained. Could I become the next
maintainer of a c2rust libyaml translation if a horrible flaw comes out
about the current version? It's not impossible, and the c2rust approach
has all kinds of benefits. But I can't _eliminate_ the possibility of
linking directly to libyaml in the future, and losing the ability to
unwind across a true FFI boundary.

One huge caveat: my current vision might actually violate Stacked
Borrows so badly that I have no choice but to give up, or reconsider it
so strongly that it no longer resembles what I have in mind right now.
Or maybe it'll work under Tree Borrows, or maybe it'll wildly violate
_any_ reasonable aliasing model for Rust. There will still be an element
of guessing and checking.