Rename xt::yaml::chonkster to xt::yaml::parser
The original point of the silly name was to serve as a placeholder until
I fully replaced the original chunker module. But now, after doing what
I thought would be a minimal and hacky integration, I can see a TON of
value in keeping `unsafe` completely out of the chunker. Its work is
already complicated enough.