Clean up xt::yaml::parser even further
A big one was to _remove_ Deref[Mut] impls from Parser. This way, it's
guaranteed that we only derefrence self.parser when we have &mut self,
so there's no risk of accidentally aliasing the pointer (especially if
we pass it back to unsafe_libyaml, and it becomes raw again, and
anything goes at that point).