Fix build errors on Linux aarch64 + armv7
Something has introduced a disturbing mismatch between the
`libc::c_char` field in the parser struct and the `std::ffi::c_char`
argument to my new function (which is exactly the same type that
`CStr::from_ptr` takes). Somehow, the former is taking the signedness of
the x86_64 build host rather than the unsignedness of the targets.

I tested the musl target in a Docker container on my aarch64 MacBook, so
AFAIK the cast here isn't destroying anything too badly.