Add a RELEASE-README note about macOS Gatekeeper
Since Sequoia got rid of the right-clicking bit (for fairly defensible
reasons), it's become far more annoying to get a browser-downloaded copy
of xt running on macOS. So, I figure this deserves an explicit mention
and a helpful command.

Ideally I would remove this section for Linux users so they never have
to think about the existence of Apple and macOS, but it's not critical
enough for me to want to template the file.