Ensure ReadState can't leak if Parser::new panics
This is quite a razor-sharp edge case, especially considering our use of
`panic = "abort"` on xt binary builds, and the fact that our panic is
supposedly only possible if we OOM (in which case, doesn't the default
global allocator abort the program itself?).

That said, the pre-refactor Chunker did this right, so I might as well
not give that up. I've also tried to be more specific about the Box
aliasing concerns that motivate this setup.