Simplify the input path iterator
I feel like the management of the "used" field was more confusing to me
after being away from this for a while. Option::take is so natural for
this case that I wish I'd thought of it before.

(I actually started with std::option::IntoIter here, before realizing
that all the iterator does is call Option::take.)