Try a few more pedantic Clippy lints
You still can't convince me that an if-else is _more_ readable than my
match on a boolean in LocatedError::from_parts. I also don't care for
the insistence on replacing closures directly with their functions;
shouldn't the compiler be able to optimize that?

I'd love to turn on the one about un-backticked items in doc comments,
but it keeps picking up on "MessagePack" despite that being the human
name of that particular data format. I'm not sure if you can customize
exclusions to that rule.