Fix a crash involving unknown YAML aliases
I finally tried fuzzing xt, and immediately discovered this issue in the
chunker. It assumed that every YAML document would contain at least one
scalar or collection between the start and end of a document, and
unwrapped an Option based on that faulty knowledge.

I haven't committed the fuzzing setup at this point, but I probably
should after releasing this.